adds views_handler_field_date for AP style date handling
To allow for system dates such as post date, update date, create date to be formatted in AP style for output, a views_handler was created. Due to the nature of the views class that is extended, a date object is required for return and not all options (those that require direct string output) that are available to the field formatter are available for the date formatter.